+++ core/src/main/scala/cats/package.scala
@@ -10,6 +10,11 @@ package object cats {
   type ⊥ = Nothing
   type ⊤ = Any
 
+  /** [[cats.Inject]][F, G] */
+  type :<:[F[_], G[_]] = Inject[F, G]
+
+  /** [[cats.Inject]][F, G] */
+  type :≺:[F[_], G[_]] = Inject[F, G]
 
 /**
  * Identity, encoded as `type Id[A] = A`, a convenient alias to make
